Add: camera and viewport resize

+++ Hazel/src/Hazel/ImGui/ImGuiLayer.cpp
@@ -77,8 +77,8 @@ namespace hazel
 	{
 		ImGuiIO& io = ImGui::GetIO();
 		auto app = Application::GetInstance();
-		io.DisplaySize = ImVec2((app->GetWindow().GetSize().x),
-			(app->GetWindow().GetSize().y));
+		io.DisplaySize = ImVec2((float)app->GetWindow().GetSize().x,
+			(float)app->GetWindow().GetSize().y);
 
 		float time = (float)glfwGetTime();
 		io.DeltaTime = m_Time > 0.0f ? (time - m_Time) : (1.0f / 60.0f);
