+++ Hazel/src/Hazel/Window.h
@@ -52,7 +52,7 @@ namespace hazel
 
 		virtual void OnUpdate() = 0;
 
-		virtual glm::vec2 GetSize()const = 0;
+		virtual glm::ivec2 GetSize()const = 0;
 
 		/*virtual uint32_t GetWidth() const = 0;
 		virtual uint32_t GetHeight() const = 0;*/
